pyproject.toml
[project]
name = "cadence"
version = "2.0.1"
description = "Multi-tenant, multi-orchestrator AI agent platform"
authors = [
    { name = "jonaskahn", email = "me@ifelse.one" }
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "cadence-sdk>=2.0.0,<3.0.0",

    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",

    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "motor>=3.3.0",
    "redis[hiredis]>=5.0.0",
    "python-ulid>=3.0.0",

    "passlib[argon2] >=0.1.10",
    "aio-pika (>=9.0.0,<10.0.0)",
    "aioboto3>=13.0",

    "litellm (>=1.81.16,<2.0.0)",

    "langchain-core>=1.2.0,<2.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    "langchain-anthropic>=1.0.0,<2.0.0",
    "langchain-google-genai>=4.0.0,<5.0.0",
    "langchain-groq>=1.0.0,<2.0.0",
    "langgraph>=1.0.0,<2.0.0",
    "langchain-litellm (>=0.5.1,<0.6.0)",

    "google-adk (>=1.26.0,<2.0.0)",

    "openai-agents (>=0.10.2,<0.11.0)"

]

[tool.poetry]
packages = [
    { include = "cadence", from = "src" },
]

[tool.poetry.scripts]
cadence-ui = "ui.run:main"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
httpx = ">=0.27.0"
ruff = ">=0.15.2"
black = ">=26.1.0"
greenlet = ">=3.3.1"
twine = "^6.2.0"

[tool.poetry.group.docs.dependencies]
zensical = ">=0.0.23"
mkdocs-git-revision-date-localized-plugin = "^1.5.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
